+++ done/pps-list-nodes.c
@@ -33,19 +33,19 @@ int main(int argc, char* argv[])
 
     error_code error_init = client_init(client_i);
 
-    if(error_init!=ERR_NONE) {
+    if(error_init != ERR_NONE) {
         printf("FAIL\n");
         return -1;
     }
 
     /* Pings every server */
-    for(int i = 0; i<client.server.size; i++) {
+    for(size_t i = 0; i < client.server.size; i++) {
         char* buffer_send = malloc(0);
         send_packet(client.socket, buffer_send, 0, client.server.nodes[i]);
 
         int error_receive = recv(client.socket, NULL,0,0);
         char* status;
-        if(error_receive==-1) {
+        if(error_receive == -1) {
             status = "FAIL";
 
         } else {
